MASSIVE reorganozation by Jacob, reflection and specilar controls. Project finished for eval. Thank you for your service, soldier.

+++ include/color.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   color.h                                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: nmonzon <nmonzon@student.42.fr>            +#+  +:+       +#+        */
+/*   By: jgraf <jgraf@student.42heilbronn.de>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/03/18 15:49:17 by jgraf             #+#    #+#             */
-/*   Updated: 2025/03/19 17:55:08 by nmonzon          ###   ########.fr       */
+/*   Updated: 2025/03/21 15:59:55 by jgraf            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,10 +17,10 @@
 
 typedef struct s_color
 {
-	uint8_t	r;
-	uint8_t	g;
-	uint8_t	b;
-	uint8_t	a;
+	int	r;
+	int	g;
+	int	b;
+	int	a;
 }	t_color;
 
 #endif
